+++ src/test-jdk17/java/com/fasterxml/jackson/dataformat/xml/records/tofix/XmlRecordDeser735Test.java
@@ -1,4 +1,4 @@
-package com.fasterxml.jackson.dataformat.xml.records.failing;
+package com.fasterxml.jackson.dataformat.xml.records.tofix;
 
 import org.junit.jupiter.api.Test;
